inference-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
requests==2.32.3
pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==1.26.4

# ML
torch==2.3.1
torchvision==0.18.1
ultralytics==8.2.103

firebase-admin==6.5.0